requirements.txt
langchain
langchain-community
ollama
langchain-ollama
wikipedia
duckduckgo-search
pydantic
python-dotenv
